Remove EDH metadata api methods & tests
